Reusing logic for reproduce video

+++ domain/src/main/java/innovappte/mobile/domain/models/Action.kt
@@ -5,5 +5,5 @@ import java.io.Serializable
 class Action: Serializable{
     var action: String? = null
     var button: String? = null
-    var index: Int = 0
+    var index: Long = 0
 }
